修改各种预警指数为整数

+++ src/main/java/com/zy/gongzhonghao/management/bean/TotalWarning.java
@@ -15,15 +15,15 @@ public class TotalWarning {
 
     private Date statisticsDate;
 
-    private Float craneWeight;
+    private Integer craneWeight;
 
-    private Float lifterWeight;
+    private Integer lifterWeight;
 
-    private Float noiseWarning;
+    private Integer noiseWarning;
 
-    private Float dustWarning;
+    private Integer dustWarning;
 
-    private Float carWarning;
+    private Integer carWarning;
 
     private Date insert_time;
